Create a new stream from a specific address

Creates a stream from a specified address (the address pays for the transaction). This is useful when the node has multiple addresses and you want to control which address appears as the creator.

Usage

mc_create_stream_from(
  conn,
  from_address,
  name,
  open = TRUE,
  custom_fields = NULL
)

Arguments

conn

A connection object created by mc_connect.

from_address

Character string. The address that will create the stream.

name

Character string. Stream name.

open

Either logical or a list of parameters (see mc_create_stream).

custom_fields

Optional list of custom fields.

Value

A character string containing the transaction ID.

See also

Examples

if (FALSE) { # \dontrun{
txid <- mc_create_stream_from(conn, "1A...", "mystream", open = TRUE)
} # }