multichainr: R Interface to the 'MultiChain' Blockchain RPC API

Provides a comprehensive R interface to the 'MultiChain' blockchain JSON-RPC API https://www.multichain.com/developers/json-rpc-api/. Allows users to manage blockchain nodes, create and subscribe to data streams, issue assets, and manage network permissions directly from the R console. Supports both local node management and remote server interaction.

See also

Author

Maintainer: Aliaksandr Martsinkevich alex@capsula.by (ORCID)